Move the inner classes/interface from ConnectionManager at top level and rename a bit
